Work on linking the nodes together.
I wanted to commit this early so I could make a not that I could save on
calculations by not running check node down if check node right failed
was a sucsess. however whith unreal only alowing one exit point from a
macro this would be more effort than the small performance benafit it
would give on starting a game.